Add two numbers as linked lists (Exercise from LeetCode)
Problem: https://leetcode.com/problems/add-two-numbers/

Revised the problem a little so that the output is an integer and not a linked list.

You are given two non-empty linked lists representing two non-negative integers. The digits are stored in reverse order and each of their nodes contain a single digit. Add the two numbers and return it as a linked list.

You may assume the two numbers do not contain any leading zero, except the number 0 itself.

Input: (2 -> 4 -> 3) + (5 -> 6 -> 4)
Output: 7 -> 0 -> 8